requirements.txt
PyPDF2
transformers
sentence-transformers
faiss-cpu
gradio
